Transfer login/register system into 'accounts' app
