+++ templates/candidates/profile.html
@@ -25,7 +25,7 @@
                 Applications</a>
             </li>
           </ul>
-          <form method="post" action="{% url 'logout' %}">
+          <form method="post" action="{% url 'accounts:logout' %}">
             {% csrf_token %}
             <button type="submit" class="btn btn-outline-danger mt-4 w-100" style="border-radius: 10px;">Log out
             </button>
