+++ templates/includes/header.html
@@ -26,7 +26,7 @@
             My Profile
           </a>
         {% else %}
-          <a href="{% url 'login' %}" class="btn btn-dark px-4" style="border-radius: 999px;">
+          <a href="{% url 'accounts:login' %}" class="btn btn-dark px-4" style="border-radius: 999px;">
             Login
           </a>
         {% endif %}
