+++ templates/registration/login.html
@@ -46,7 +46,7 @@
 
   <p class="text-center small-link mb-0">
     Don’t have an account?
-    <a href="{% url 'register' %}">Register</a>
+    <a href="{% url 'accounts:register' %}">Register</a>
   </p>
 </div>
