+++ templates/registration/signup.html
@@ -76,7 +76,7 @@
 
     <p class="text-center small-link mb-0">
       Already have an account?
-      <a href="{% url 'login' %}">Login</a>
+      <a href="{% url 'accounts:login' %}">Login</a>
     </p>
   </div>
